Module libprosic::model::priors::tumor_normal

Structs

TumorNormalModel
Tumor-normal prior model using ploidy, heterozygosity (in normal tissue) and tumor mutation rate per effective cell division. The latter is the quotient mu/beta, with mu being the mutation rate and beta being the fraction of effective cell divisions (both lineages survive). Alone, the parameters are not observable. However, mu/beta can be estimated from e.g. SNV calls. It is the slope of the linear model y = mu/beta * (x - 1 / fmax), withx` being the reciprocal of the observed allele frequencies and y being the number of observed mutations corresponding to each frequency (see Williams et al. Nature Genetics 2016).