Struct libprosic::model::sample::RecordBuffer

pub struct RecordBuffer {
    // some fields omitted
}

Ringbuffer of BAM records. This data structure ensures that no bam record is read twice while extracting observations for given variants.

Methods

impl RecordBuffer

fn new(bam: IndexedReader, window: u32, use_secondary: bool) -> Self

Create a new RecordBuffer.

fn fill(&mut self, chrom: &[u8], start: u32, end: u32) -> Result<(), Box<Error>>

Fill buffer around the given interval.

fn iter(&self) -> Iter<Record>