Struct libprosic::model::sample::Sample

pub struct Sample {
    // some fields omitted
}

A sequenced sample, e.g., a tumor or a normal sample.

Methods

impl Sample

fn new(bam: IndexedReader, pileup_window: u32, use_fragment_evidence: bool, use_secondary: bool, use_mapq: bool, insert_size: InsertSize, likelihood_model: LatentVariableModel, prob_spurious_isize: Prob, prob_missed_insertion_alignment: Prob, prob_missed_deletion_alignment: Prob, prob_spurious_indel_alignment: Prob) -> Self

Create a new Sample.

Arguments

bam - BAM file with the aligned and deduplicated sequence reads.

pileup_window - Window around the variant that shall be search for evidence (e.g. 5000).

use_fragment_evidence - Whether to use read pairs that are left and right of variant.

use_secondary - Whether to use secondary alignments.

insert_size - estimated insert size

prior_model - Prior assumptions about allele frequency spectrum of this sample.

likelihood_model - Latent variable model to calculate likelihoods of given observations.

prob_spurious_isize - rate of wrongly reported insert size abberations (mapper dependent, BWA: 0.01332338, LASER: 0.05922201)

prob_missed_insertion_alignment - rate of missed insertion alignments if insertion is present (mapper dependent, BWA: 0.2138, LASER: 0.3460)

prob_missed_deletion_alignment - rate of missed deletion alignments if deletion is present (mapper dependent, BWA: 0.0310, LASER: 0.0964)

prob_spurious_indel_alignment - rate of wrongly reported indel alignments if indel is absent (mapper dependent)

fn max_indel_dist(self, dist: u32) -> Self

fn max_indel_len_diff(self, diff: u32) -> Self

fn likelihood_model(&self) -> LatentVariableModel

Return likelihood model.

fn extract_observations(&mut self, chrom: &[u8], start: u32, variant: Variant) -> Result<Vec<Observation>, Box<Error>>

Extract observations for the given variant.